pyproject.toml
[tool.poetry]
name = "agents"
version = "0.1.0"
description = "an attempt at orchestrating different agents"
authors = ["sabarish"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
openai = "^1.59.4"
numpy = "^2.2.1"
pandas = "^2.2.3"
faiss-cpu = "^1.9.0.post1"
pyyaml = "^6.0.2"
requests = "^2.32.3"
tiktoken = "^0.8.0"
streamlit = "^1.41.1"
langchain = "^0.3.14"
ipykernel = "^6.29.5"
tabulate = "^0.9.0"
openpyxl = "^3.1.5"
autogen-agentchat = ">=0.2,<1.0"
autogen-ext = "^0.4.0"
duckdb = "^1.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
